+++ app/src/main/assets/changelog.md
@@ -1,3 +1,9 @@
+### 1.1.1
+* **Fix:** Hide hash calculations for crypt remotes
+* **Fix:** Crash when rclone fails
+
+***
+
 ### 1.1.0
 * **New:** Support password protected configurations
 * **New:** Calculate file MD5 & SHA1